Add support for vulkan semantics
Invoke glslangValidator by default with -V so that it
works for vulkan shaders. This is not well tested and
currently only works on linux (due to using /dev/null for
the binary output).
This might decrase performance (since a spirv binary is
generated) and might break some default glsl shaders since
it always uses vulkan semantics.